FAC-361: Remove CheckSpaceExistsPort and move checks to adapter

+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/adapter/out/persistence/space/SpacePersistenceJpaAdapter.java
@@ -95,6 +95,8 @@ public class SpacePersistenceJpaAdapter implements
 
     @Override
     public void deleteById(long spaceId) {
+        if (!repository.existsByIdAndDeletedFalse(spaceId))
+            throw new ResourceNotFoundException(SPACE_ID_NOT_FOUND);
         repository.delete(spaceId);
     }
 }
